Initial uploading of images removed. On every package creation the image (if exists) is uploaded (if needed).
